v2: merge main into v2 again (#2937)

+++ contrib/graphql-go/graphql/appsec_test.go
@@ -140,7 +140,7 @@ func TestAppSec(t *testing.T) {
 				require.NotEmpty(t, spans)
 				// The last finished span (which is GraphQL entry) should have the "_dd.appsec.enabled" tag.
 				span := spans[len(spans)-1]
-				require.Equal(t, 1, span.Tag("_dd.appsec.enabled"))
+				require.Equal(t, float64(1), span.Tag("_dd.appsec.enabled"))
 				type ddAppsecJSON struct {
 					Triggers []struct {
 						Rule struct {
